+++ packages/theme-chalk/icons.js
@@ -4,13 +4,18 @@ import Airplane from '../../packages/theme-chalk/src/icons/airplane.svg';
 import ArrowBack from '../../packages/theme-chalk/src/icons/arrow-back.svg';
 import ArrowBoth from '../../packages/theme-chalk/src/icons/arrow-both.svg';
 import ArrowDownLong from '../../packages/theme-chalk/src/icons/arrow-down-long.svg';
+import ArrowDownSoft from '../../packages/theme-chalk/src/icons/arrow-down-soft.svg';
 import ArrowDown from '../../packages/theme-chalk/src/icons/arrow-down.svg';
+import ArrowLeftLong from '../../packages/theme-chalk/src/icons/arrow-left-long.svg';
 import ArrowLeft from '../../packages/theme-chalk/src/icons/arrow-left.svg';
 import ArrowNext from '../../packages/theme-chalk/src/icons/arrow-next.svg';
+import ArrowRightLong from '../../packages/theme-chalk/src/icons/arrow-right-long.svg';
 import ArrowRight from '../../packages/theme-chalk/src/icons/arrow-right.svg';
+import ArrowUpSoft from '../../packages/theme-chalk/src/icons/arrow-up-soft.svg';
 import ArrowsReverse from '../../packages/theme-chalk/src/icons/arrows-reverse.svg';
 import Bath from '../../packages/theme-chalk/src/icons/bath.svg';
 import Bell from '../../packages/theme-chalk/src/icons/bell.svg';
+import Boiler from '../../packages/theme-chalk/src/icons/boiler.svg';
 import Book from '../../packages/theme-chalk/src/icons/book.svg';
 import Bucket from '../../packages/theme-chalk/src/icons/bucket.svg';
 import Calendar from '../../packages/theme-chalk/src/icons/calendar.svg';
@@ -21,11 +26,15 @@ import Check from '../../packages/theme-chalk/src/icons/check.svg';
 import CircleAnglesSquare from '../../packages/theme-chalk/src/icons/circle-angles-square.svg';
 import CirclePlus from '../../packages/theme-chalk/src/icons/circle-plus.svg';
 import City from '../../packages/theme-chalk/src/icons/city.svg';
+import Conductor from '../../packages/theme-chalk/src/icons/conductor.svg';
+import Cooler from '../../packages/theme-chalk/src/icons/cooler.svg';
 import Copy from '../../packages/theme-chalk/src/icons/copy.svg';
 import CreditCard from '../../packages/theme-chalk/src/icons/credit-card.svg';
 import CrossBold from '../../packages/theme-chalk/src/icons/cross-bold.svg';
 import Cross from '../../packages/theme-chalk/src/icons/cross.svg';
+import CupHot from '../../packages/theme-chalk/src/icons/cup-hot.svg';
 import Cup from '../../packages/theme-chalk/src/icons/cup.svg';
+import CustomsOfficer from '../../packages/theme-chalk/src/icons/customs-officer.svg';
 import DisabledPerson from '../../packages/theme-chalk/src/icons/disabled-person.svg';
 import DotsHorizontal from '../../packages/theme-chalk/src/icons/dots-horizontal.svg';
 import DotsVertical from '../../packages/theme-chalk/src/icons/dots-vertical.svg';
@@ -33,24 +42,31 @@ import DoubleBed from '../../packages/theme-chalk/src/icons/double-bed.svg';
 import Download from '../../packages/theme-chalk/src/icons/download.svg';
 import Dryer from '../../packages/theme-chalk/src/icons/dryer.svg';
 import Error from '../../packages/theme-chalk/src/icons/error.svg';
+import Extinguisher from '../../packages/theme-chalk/src/icons/extinguisher.svg';
 import Files from '../../packages/theme-chalk/src/icons/files.svg';
 import FlightArrive from '../../packages/theme-chalk/src/icons/flight-arrive.svg';
 import FlightDepart from '../../packages/theme-chalk/src/icons/flight-depart.svg';
 import FullStar from '../../packages/theme-chalk/src/icons/full-star.svg';
+import GearNWrench from '../../packages/theme-chalk/src/icons/gear-n-wrench.svg';
+import Glass from '../../packages/theme-chalk/src/icons/glass.svg';
 import Gym from '../../packages/theme-chalk/src/icons/gym.svg';
 import HalfStar from '../../packages/theme-chalk/src/icons/half-star.svg';
+import Hanger from '../../packages/theme-chalk/src/icons/hanger.svg';
 import Hotel from '../../packages/theme-chalk/src/icons/hotel.svg';
 import Insurance from '../../packages/theme-chalk/src/icons/insurance.svg';
 import Iron from '../../packages/theme-chalk/src/icons/iron.svg';
 import ItemList from '../../packages/theme-chalk/src/icons/item-list.svg';
 import ItemSort from '../../packages/theme-chalk/src/icons/item-sort.svg';
+import KnifeFork2 from '../../packages/theme-chalk/src/icons/knife-fork-2.svg';
 import KnifeFork from '../../packages/theme-chalk/src/icons/knife-fork.svg';
 import Label from '../../packages/theme-chalk/src/icons/label.svg';
 import Loader from '../../packages/theme-chalk/src/icons/loader.svg';
 import Location from '../../packages/theme-chalk/src/icons/location.svg';
+import Luggage from '../../packages/theme-chalk/src/icons/luggage.svg';
 import ManWoman from '../../packages/theme-chalk/src/icons/man-woman.svg';
 import Message from '../../packages/theme-chalk/src/icons/message.svg';
 import Minus from '../../packages/theme-chalk/src/icons/minus.svg';
+import Negotiations from '../../packages/theme-chalk/src/icons/negotiations.svg';
 import NoSmoking from '../../packages/theme-chalk/src/icons/no-smoking.svg';
 import Parking from '../../packages/theme-chalk/src/icons/parking.svg';
 import Pencil from '../../packages/theme-chalk/src/icons/pencil.svg';
@@ -60,15 +76,23 @@ import Phone from '../../packages/theme-chalk/src/icons/phone.svg';
 import Pillows from '../../packages/theme-chalk/src/icons/pillows.svg';
 import Plus from '../../packages/theme-chalk/src/icons/plus.svg';
 import Pocket from '../../packages/theme-chalk/src/icons/pocket.svg';
+import Pram from '../../packages/theme-chalk/src/icons/pram.svg';
+import Ramp from '../../packages/theme-chalk/src/icons/ramp.svg';
 import Repeat from '../../packages/theme-chalk/src/icons/repeat.svg';
 import SearchSmall from '../../packages/theme-chalk/src/icons/search-small.svg';
 import Search from '../../packages/theme-chalk/src/icons/search.svg';
+import SeatPerson from '../../packages/theme-chalk/src/icons/seat-person.svg';
+import Service from '../../packages/theme-chalk/src/icons/service.svg';
+import Shower from '../../packages/theme-chalk/src/icons/shower.svg';
 import SingleBed from '../../packages/theme-chalk/src/icons/single-bed.svg';
+import Smoke from '../../packages/theme-chalk/src/icons/smoke.svg';
 import Snow2 from '../../packages/theme-chalk/src/icons/snow-2.svg';
 import Snow from '../../packages/theme-chalk/src/icons/snow.svg';
+import Sofa from '../../packages/theme-chalk/src/icons/sofa.svg';
 import Star from '../../packages/theme-chalk/src/icons/star.svg';
 import StatsPie from '../../packages/theme-chalk/src/icons/stats-pie.svg';
 import Success from '../../packages/theme-chalk/src/icons/success.svg';
+import Teddybear from '../../packages/theme-chalk/src/icons/teddybear.svg';
 import Ticket from '../../packages/theme-chalk/src/icons/ticket.svg';
 import Time from '../../packages/theme-chalk/src/icons/time.svg';
 import Train from '../../packages/theme-chalk/src/icons/train.svg';
@@ -77,8 +101,10 @@ import Tv from '../../packages/theme-chalk/src/icons/tv.svg';
 import UmbrellaAngle from '../../packages/theme-chalk/src/icons/umbrella-angle.svg';
 import Umbrella from '../../packages/theme-chalk/src/icons/umbrella.svg';
 import Upload from '../../packages/theme-chalk/src/icons/upload.svg';
+import Utility from '../../packages/theme-chalk/src/icons/utility.svg';
 import Vault from '../../packages/theme-chalk/src/icons/vault.svg';
 import Warning from '../../packages/theme-chalk/src/icons/warning.svg';
+import Wc from '../../packages/theme-chalk/src/icons/wc.svg';
 import Wifi from '../../packages/theme-chalk/src/icons/wifi.svg';
 import Window from '../../packages/theme-chalk/src/icons/window.svg';
 
@@ -89,13 +115,18 @@ module.exports = {
   ArrowBack,
   ArrowBoth,
   ArrowDownLong,
+  ArrowDownSoft,
   ArrowDown,
+  ArrowLeftLong,
   ArrowLeft,
   ArrowNext,
+  ArrowRightLong,
   ArrowRight,
+  ArrowUpSoft,
   ArrowsReverse,
   Bath,
   Bell,
+  Boiler,
   Book,
   Bucket,
   Calendar,
@@ -106,11 +137,15 @@ module.exports = {
   CircleAnglesSquare,
   CirclePlus,
   City,
+  Conductor,
+  Cooler,
   Copy,
   CreditCard,
   CrossBold,
   Cross,
+  CupHot,
   Cup,
+  CustomsOfficer,
   DisabledPerson,
   DotsHorizontal,
   DotsVertical,
@@ -118,24 +153,31 @@ module.exports = {
   Download,
   Dryer,
   Error,
+  Extinguisher,
   Files,
   FlightArrive,
   FlightDepart,
   FullStar,
+  GearNWrench,
+  Glass,
   Gym,
   HalfStar,
+  Hanger,
   Hotel,
   Insurance,
   Iron,
   ItemList,
   ItemSort,
+  KnifeFork2,
   KnifeFork,
   Label,
   Loader,
   Location,
+  Luggage,
   ManWoman,
   Message,
   Minus,
+  Negotiations,
   NoSmoking,
   Parking,
   Pencil,
@@ -145,15 +187,23 @@ module.exports = {
   Pillows,
   Plus,
   Pocket,
+  Pram,
+  Ramp,
   Repeat,
   SearchSmall,
   Search,
+  SeatPerson,
+  Service,
+  Shower,
   SingleBed,
+  Smoke,
   Snow2,
   Snow,
+  Sofa,
   Star,
   StatsPie,
   Success,
+  Teddybear,
   Ticket,
   Time,
   Train,
@@ -162,8 +212,10 @@ module.exports = {
   UmbrellaAngle,
   Umbrella,
   Upload,
+  Utility,
   Vault,
   Warning,
+  Wc,
   Wifi,
   Window
 };